Get rid of `state` altogether.
Having a `state` makes no sense in our context, it was introduced for
some obscure reasons at the dawn of the project. We ended up almost
never using it, except for some contrived use-case where we stored the
eigenpairs of the Laplacian there.

This does introduce breaking changes, if one is not using the wrappers
a la GpytorchGeometricKernel.

This commit also does not deal with the example notebooks, so most of
them will be broken to be fixed later.